+++ bulletin_board/server/app/assets/config/manifest.js
@@ -18,3 +18,6 @@
 
 //= link sandbox/tally.js
 //= link sandbox/tally.css
+
+//= link sandbox/elections/new.css
+//= link sandbox/elections/new.js